refactor(statusTag): consolidate and enhance status tag component functionality
- Merged the status tag logic into a single file, removing the deprecated statusTag.tsx.
- Introduced a new StatusTagWrapper component to manage tooltips and event bindings more effectively.
- Enhanced rendering logic to ensure proper handling of status values and improved tooltip visibility.
- Updated utility functions for better normalization of status options and legacy mappings.
- Improved styling and layout for consistent user experience across different scenarios.
